+++ lib/calendrical/calendars/lunar_japanese.ex
@@ -835,20 +835,20 @@ defmodule Calendrical.LunarJapanese do
   ### Examples
 
       # Y1381 (= AD 2025) has an intercalary 6th month (閏6月)
-      iex> Calendrical.LunarJapanese.ordinal_month(1381, {6, :leap})
+      iex> Calendrical.LunarJapanese.ordinal_month_from_traditional(1381, {6, :leap})
       {:ok, 7}
 
-      iex> Calendrical.LunarJapanese.ordinal_month(1381, 7)
+      iex> Calendrical.LunarJapanese.ordinal_month_from_traditional(1381, 7)
       {:ok, 8}
 
-      iex> Calendrical.LunarJapanese.ordinal_month(1380, {6, :leap})
+      iex> Calendrical.LunarJapanese.ordinal_month_from_traditional(1380, {6, :leap})
       {:error, :invalid_leap_month}
 
   """
-  @spec ordinal_month(Calendar.year(), Lunisolar.lunar_month()) ::
+  @spec ordinal_month_from_traditional(Calendar.year(), Lunisolar.lunar_month()) ::
           {:ok, Calendar.month()} | {:error, :invalid_month | :invalid_leap_month}
-  def ordinal_month(year, lunar_month) do
-    Lunisolar.ordinal_month(year, lunar_month, epoch(), &location/1)
+  def ordinal_month_from_traditional(year, lunar_month) do
+    Lunisolar.ordinal_month_from_traditional(year, lunar_month, epoch(), &location/1)
   end
 
   @doc """